add basic user controller
