+++ server/src/gym/dto/GymDto.ts
@@ -1,0 +1,5 @@
+export default class GymDto {
+  constructor(
+    public name: string
+  ){}
+}
