+++ server/src/gym/gymController.ts
@@ -1,7 +1,8 @@
 import {Request, Response} from "express";
-import { JsonController, Get, Req, Res } from "routing-controllers";
+import { JsonController, Get, Req, Res, UseBefore } from "routing-controllers";
 import GymProvider from "./gymProvider";
 import { Service } from "typedi";
+import passport = require("passport");
 
 @Service()
 @JsonController()
@@ -9,7 +10,9 @@ export class GymController {
   constructor(private gymProvider: GymProvider) {}
 
   @Get("/gyms")
+  @UseBefore(passport.authenticate('jwt', { session: false }))
   public async getAll(@Req() request: Request, @Res() response: Response) {
-    return response.json(this.gymProvider.getAll());
+    const allGymsResponse = await this.gymProvider.getAll();
+    return response.json(allGymsResponse.gyms);
   }
 }
