+++ server/src/gym/responses/getAllGymsResponse.ts
@@ -1,0 +1,5 @@
+import GymDto from "../dto/GymDto";
+
+export default class GetAllGymsResponse {
+  constructor(public gyms: GymDto[]) {}
+}
