--- server/src/services/gym/gymProvider.ts
@@ -1,12 +1,0 @@
-import { Service } from 'typedi';
-import { Gym } from "../../models/gym/Gym";
-
-@Service()
-export default class GymProvider {
-  constructor() {}
-  
-  public async getAll() {
-    const gyms = await Gym.findAll();
-    return gyms;
-  }
-}
