--- server/src/services/index.ts
@@ -1,4 +1,0 @@
-// import gymRoutes from "./gym/routes";
-import userRoutes from "./user/routes";
-
-export default [...userRoutes];
