--- server/src/utils/index.ts
@@ -1,31 +1,0 @@
-import { Router, Request, Response, NextFunction } from "express";
-
-type Wrapper = ((router: Router) => void);
-
-export const applyMiddleware = (
-    middleware: Wrapper[],
-    router: Router
-) => {
-    for (const f of middleware) {
-        f(router);
-    }
-};
-
-type Handler = (
-    req: Request,
-    res: Response,
-    next: NextFunction
-) => Promise<void> | void;
-
-type Route = {
-    path: string;
-    method: string;
-    handler: Handler | Handler[];
-};
-
-export const applyRoutes = (routes: Route[], router: Router) => {
-    for (const route of routes) {
-        const { method, path, handler } = route;
-        (router as any)[method](path, handler);
-    }
-};
